+++ tests/compat/py39.py
@@ -1,0 +1,9 @@
+from jaraco.test.cpython import from_test_support, try_import
+
+
+os_helper = try_import('os_helper') or from_test_support(
+    'FS_NONASCII', 'skip_unless_symlink', 'temp_dir'
+)
+import_helper = try_import('import_helper') or from_test_support(
+    'modules_setup', 'modules_cleanup'
+)