+++ tests/compat/test_py39_compat.py
@@ -2,7 +2,7 @@ import sys
 import pathlib
 import unittest
 
-from . import fixtures
+from .. import fixtures
 from importlib_metadata import (
     distribution,
     distributions,
@@ -14,8 +14,7 @@ from importlib_metadata import (
 
 class OldStdlibFinderTests(fixtures.DistInfoPkgOffPath, unittest.TestCase):
     def setUp(self):
-        python_version = sys.version_info[:2]
-        if python_version < (3, 8) or python_version > (3, 9):
+        if sys.version_info >= (3, 10):
             self.skipTest("Tests specific for Python 3.8/3.9")
         super().setUp()